+++ hera/debugger/minilanguage.py
@@ -50,10 +50,6 @@ class MiniParser:
             return RegisterNode(tkn[1])
         elif tkn[0] == TOKEN_SYMBOL:
             return SymbolNode(tkn[1])
-        elif tkn[0] == TOKEN_TRUE:
-            return BoolNode(True)
-        elif tkn[0] == TOKEN_FALSE:
-            return BoolNode(False)
         else:
             self.raise_unexpected(tkn)
 
@@ -74,7 +70,6 @@ class MiniParser:
 MemoryNode = namedtuple("MemoryNode", ["address"])
 RegisterNode = namedtuple("RegisterNode", ["value"])
 IntNode = namedtuple("IntNode", ["value"])
-BoolNode = namedtuple("BoolNode", ["value"])
 SymbolNode = namedtuple("SymbolNode", ["value"])
 MinusNode = namedtuple("MinusNode", ["arg"])
 
@@ -111,11 +106,6 @@ class MiniLexer:
             return self.advance_and_return(TOKEN_INT, length=length)
         elif ch == "-":
             return self.advance_and_return(TOKEN_MINUS)
-        elif ch == "#" and self.peek() in "tfTF":
-            if self.peek() in "tT":
-                return self.advance_and_return(TOKEN_TRUE, length=2)
-            else:
-                return self.advance_and_return(TOKEN_FALSE, length=2)
         else:
             return self.advance_and_return(TOKEN_UNKNOWN)
 
@@ -184,11 +174,6 @@ TOKEN_MEM = "TOKEN_MEM"
 TOKEN_REGISTER = "TOKEN_REGISTER"
 TOKEN_RBRACKET = "TOKEN_RBRACKET"
 TOKEN_SYMBOL = "TOKEN_SYMBOL"
-TOKEN_TRUE = "TOKEN_TRUE"
-TOKEN_FALSE = "TOKEN_FALSE"
 TOKEN_MINUS = "TOKEN_MINUS"
 TOKEN_EOF = "TOKEN_EOF"
 TOKEN_UNKNOWN = "TOKEN_UNKNOWN"
-
-
-FLAG_LITERALS = ("f_cb", "f_c", "f_v", "f_z", "f_s")