+++ test/test_unit/test_minilanguage.py
@@ -1,5 +1,4 @@
 from hera.debugger.minilanguage import (
-    BoolNode,
     IntNode,
     MemoryNode,
     MiniLexer,
@@ -8,14 +7,12 @@ from hera.debugger.minilanguage import (
     RegisterNode,
     SymbolNode,
     TOKEN_EOF,
-    TOKEN_FALSE,
     TOKEN_INT,
     TOKEN_MEM,
     TOKEN_MINUS,
     TOKEN_RBRACKET,
     TOKEN_REGISTER,
     TOKEN_SYMBOL,
-    TOKEN_TRUE,
     TOKEN_UNKNOWN,
 )
 
@@ -30,7 +27,7 @@ def test_lex_mini_language_with_small_example():
 
 def test_lex_mini_language_with_big_example():
     # This isn't a syntactically valid expression, but it doesn't matter to the lexer.
-    lexer = MiniLexer("M[FP_alt] R15 0xabc some_symbol #t #F -10 ?")
+    lexer = MiniLexer("M[FP_alt] R15 0xabc some_symbol -10 ?")
 
     assert lexer.next_token() == (TOKEN_MEM, "M[")
     assert lexer.next_token() == (TOKEN_REGISTER, "FP_alt")
@@ -38,8 +35,6 @@ def test_lex_mini_language_with_big_example():
     assert lexer.next_token() == (TOKEN_REGISTER, "R15")
     assert lexer.next_token() == (TOKEN_INT, "0xabc")
     assert lexer.next_token() == (TOKEN_SYMBOL, "some_symbol")
-    assert lexer.next_token() == (TOKEN_TRUE, "#t")
-    assert lexer.next_token() == (TOKEN_FALSE, "#F")
     assert lexer.next_token() == (TOKEN_MINUS, "-")
     assert lexer.next_token() == (TOKEN_INT, "10")
     assert lexer.next_token() == (TOKEN_UNKNOWN, "?")
@@ -76,15 +71,6 @@ def test_parse_memory_expression_with_integer():
     assert isinstance(tree.address, MinusNode)
     assert isinstance(tree.address.arg, IntNode)
     assert tree.address.arg.value == 0o12
-
-
-def test_parse_boolean():
-    parser = MiniParser(MiniLexer("#t"))
-
-    tree = parser.parse()
-
-    assert isinstance(tree, BoolNode)
-    assert tree.value is True
 
 
 def test_parse_memory_expression_with_symbol():